# internal/dto/response/user.go
package response

import (
	"github.com/clemilsonazevedo/blog/internal/domain/enums"
	"go.bryk.io/pkg/ulid"
)

type UserLogin struct {
	Token   string `json:"token"`
	Expires int64  `json:"exp"`
}

type UserByID struct {
	ID       ulid.ULID  `json:"id" swaggertype:"string" example:"01ARZ3NDEKTSV4RRFFQ69G5FAV"`
	UserName string     `json:"username"`
	Email    string     `json:"email"`
	Role     enums.Role `json:"role" binding:"required,oneof=anonymous reader author" swaggertype:"string" enums:"anonymous,reader,author"`
}

type UserLogout struct {
	Message string `json:"message"`
}